Some minor changes

+++ src/main/java/com/infosec/accessanalysis/dal/repository/DepartmentRepositoryMSSQL.java
@@ -25,8 +25,9 @@ public class DepartmentRepositoryMSSQL implements DepartmentRepository {
                 ResultSet rs = st.executeQuery("SELECT DEPARTMENT_ID, PARENT_ID, NAME FROM dbo.DEPARTMENT WHERE IS_DELETED = 0 ORDER BY DEPARTMENT_ID");
         ) {
             while (rs.next()) {
-                Department dep = new Department(rs.getInt(1),
-                        rs.getInt(2),
+                Department dep = new Department(
+                        rs.getLong(1),
+                        rs.getLong(2),
                         rs.getString(3));
                 loadChildObjects(dep);
                 departments.add(dep);
@@ -48,8 +49,9 @@ public class DepartmentRepositoryMSSQL implements DepartmentRepository {
             st.setLong(2, from + 1 + count);
             try (ResultSet rs = st.executeQuery()) {
                 while (rs.next()) {
-                    Department dep = new Department(rs.getInt(1),
-                            rs.getInt(2),
+                    Department dep = new Department(
+                            rs.getLong(1),
+                            rs.getLong(2),
                             rs.getString(3));
                     loadChildObjects(dep);
                     departments.add(dep);
@@ -75,8 +77,9 @@ public class DepartmentRepositoryMSSQL implements DepartmentRepository {
             st.setLong(1, id);
             try (ResultSet rs = st.executeQuery()) {
                 while (rs.next()) {
-                    Department dep = new Department(rs.getInt(1),
-                            rs.getInt(2),
+                    Department dep = new Department(
+                            rs.getLong(1),
+                            rs.getLong(2),
                             rs.getString(3));
                     loadChildObjects(dep);
                     departments.add(dep);
@@ -91,13 +94,15 @@ public class DepartmentRepositoryMSSQL implements DepartmentRepository {
     public Department findParent(long id) throws SQLException {
         try (
                 Connection conn = DriverManager.getConnection(url);
+                //TODO: incorrect sql request
                 PreparedStatement st = conn.prepareStatement("SELECT DEPARTMENT_ID, PARENT_ID, NAME FROM dbo.DEPARTMENT WHERE PARENT_ID = ? AND IS_DELETED = 0");
         ) {
             st.setLong(1, id);
             try (ResultSet rs = st.executeQuery()) {
                 if (rs.next()) {
-                    Department dep = new Department(rs.getInt(1),
-                            rs.getInt(2),
+                    Department dep = new Department(
+                            rs.getLong(1),
+                            rs.getLong(2),
                             rs.getString(3));
                     loadChildObjects(dep);
                     return dep;
@@ -118,8 +123,9 @@ public class DepartmentRepositoryMSSQL implements DepartmentRepository {
                 ResultSet rs = st.executeQuery("SELECT DEPARTMENT_ID, PARENT_ID, NAME FROM dbo.DEPARTMENT WHERE PARENT_ID IS NULL AND IS_DELETED = 0");
         ) {
             while (rs.next()) {
-                Department dep = new Department(rs.getInt(1),
-                        rs.getInt(2),
+                Department dep = new Department(
+                        rs.getLong(1),
+                        rs.getLong(2),
                         rs.getString(3));
                 loadChildObjects(dep);
                 departments.add(dep);
@@ -138,8 +144,9 @@ public class DepartmentRepositoryMSSQL implements DepartmentRepository {
             st.setLong(1, id);
             try (ResultSet rs = st.executeQuery()) {
                 if (rs.next()) {
-                    Department dep = new Department(rs.getInt(1),
-                            rs.getInt(2),
+                    Department dep = new Department(
+                            rs.getLong(1),
+                            rs.getLong(2),
                             rs.getString(3));
                     loadChildObjects(dep);
                     return dep;
@@ -179,7 +186,8 @@ public class DepartmentRepositoryMSSQL implements DepartmentRepository {
             try (ResultSet rs = st.executeQuery()) {
                 List<Personage> employees = department.getEmployees();
                 while (rs.next()) {
-                    Personage personage = new Personage(rs.getInt(1),
+                    Personage personage = new Personage(
+                            rs.getLong(1),
                             rs.getString(2),
                             rs.getString(3),
                             rs.getString(4),
